Rename modules in fixtures to reflect their intent
