+++ Fixtures/ClangModules/ModuleMapGenerationCases/Package.swift
@@ -3,5 +3,5 @@ import PackageDescription
 let package = Package(
     name: "ModuleMapGenerationCases",
     targets: [
-		Target(name: "Baz", dependencies: ["Foo", "Bar", "Jaz"])]
+		Target(name: "Baz", dependencies: ["FlatInclude", "UmbrellaHeader", "UmbellaModuleNameInclude"])]
 )
