+++ Fixtures/ClangModules/ModuleMapGenerationCases/Sources/UmbrellaHeader/UmbrellaHeader.c
@@ -1,4 +1,4 @@
-#include "include/Foo/Foo.h"
+#include "include/UmbrellaHeader/UmbrellaHeader.h"
 
 int foo() {
     int a = 5;
